Deposit and withdrawal exceptions

+++ src/main/java/com/banking/banking/service/DepositService.java
@@ -1,5 +1,6 @@
 package com.banking.banking.service;
 
+import com.banking.banking.exception.ResourceNotFoundException;
 import com.banking.banking.model.Customer;
 import com.banking.banking.model.Deposit;
 import com.banking.banking.repository.*;
@@ -8,6 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
 import java.util.Optional;
 
 
@@ -25,19 +27,29 @@ public class DepositService {
 
     public ResponseEntity<Iterable<Deposit>> getAllDeposits() {
         Iterable<Deposit> deposits = depositRepository.findAll();
+
         return new ResponseEntity<>(deposits, HttpStatus.OK);
     }
 
-    public ResponseEntity<?> getDepositById(Long depositId) {
+    public ResponseEntity<?> getDepositById(Long depositId) throws ResourceNotFoundException {
         Deposit deposit = depositRepository.findById(depositId).orElse(null);
+        if (deposit == null){
+            throw new ResourceNotFoundException("Error getting deposit by id");
+        }
         return new ResponseEntity<>(deposit, HttpStatus.OK);
     }
 
-    public void updateDeposit(Deposit deposit, Long depositId) {
+    public void updateDeposit(Deposit deposit, Long depositId) throws ResourceNotFoundException{
         Deposit c = depositRepository.save(deposit);
+        if (depositId == null){
+            throw new ResourceNotFoundException("Error updating deposit");
+        }
     }
 
-    public ResponseEntity<?> deleteDepositById(Long depositId) {
+    public ResponseEntity<?> deleteDepositById(Long depositId) throws ResourceNotFoundException {
+        if (depositRepository.findById(depositId).isEmpty()){
+            throw new ResourceNotFoundException("Error deleting deposit");
+        }
         depositRepository.deleteById(depositId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
